requirements.txt
openai
python-dotenv
numpy
requests
matplotlib
plotly
scipy
scikit-learn
tenacity
pandas